Renaming ProvidersPactStates to PactReferences
