+++ PactDemo.sln
@@ -25,8 +25,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AppHost", "AspireProjects\A
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceDefaults", "AspireProjects\ServiceDefaults\ServiceDefaults.csproj", "{FD855778-252A-48F7-998E-D8B2B91004C0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProvidersPactStates", "test\ProvidersPactStates\ProvidersPactStates.csproj", "{8EB047FC-5994-44E1-B50F-C27EC0B4111E}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{B7D87FB5-427D-4FE4-94A6-E35B5FEA1877}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoConfigurations", "src\Shared\DemoConfigurations\DemoConfigurations.csproj", "{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62}"
@@ -36,6 +34,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CityService", "src\CityService\CityService.csproj", "{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CityServiceContractTests", "test\CityServiceContractTests\CityServiceContractTests.csproj", "{D927FB96-2290-4037-8EDB-5D9CA3919169}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PactReferences", "test\PactReferences\PactReferences.csproj", "{9841C637-B2C4-4927-A6AB-B8CA6869505C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,10 +59,6 @@ Global
 		{FD855778-252A-48F7-998E-D8B2B91004C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FD855778-252A-48F7-998E-D8B2B91004C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FD855778-252A-48F7-998E-D8B2B91004C0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8EB047FC-5994-44E1-B50F-C27EC0B4111E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8EB047FC-5994-44E1-B50F-C27EC0B4111E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8EB047FC-5994-44E1-B50F-C27EC0B4111E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8EB047FC-5994-44E1-B50F-C27EC0B4111E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -79,6 +75,10 @@ Global
 		{D927FB96-2290-4037-8EDB-5D9CA3919169}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D927FB96-2290-4037-8EDB-5D9CA3919169}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D927FB96-2290-4037-8EDB-5D9CA3919169}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9841C637-B2C4-4927-A6AB-B8CA6869505C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9841C637-B2C4-4927-A6AB-B8CA6869505C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9841C637-B2C4-4927-A6AB-B8CA6869505C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9841C637-B2C4-4927-A6AB-B8CA6869505C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -88,12 +88,12 @@ Global
 		{8823BF0B-D9C3-4359-B9A6-BA620D6EEB28} = {4D21B2A1-8B90-48D3-9D49-7C48B81AE481}
 		{F437E33F-5FD8-42C1-8DE9-59591D2C077F} = {6521E5A4-D661-43C7-A837-BAF2BC7048BA}
 		{FD855778-252A-48F7-998E-D8B2B91004C0} = {6521E5A4-D661-43C7-A837-BAF2BC7048BA}
-		{8EB047FC-5994-44E1-B50F-C27EC0B4111E} = {4D21B2A1-8B90-48D3-9D49-7C48B81AE481}
 		{B7D87FB5-427D-4FE4-94A6-E35B5FEA1877} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 		{EE36A974-48E9-47B1-9BA0-F9C7C74ACC62} = {B7D87FB5-427D-4FE4-94A6-E35B5FEA1877}
 		{B59B941A-69DA-4660-BF8A-C9B61B80B35C} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 		{8A5B9DAD-DC4A-4E16-AB44-71D67AE6C776} = {575244E4-337F-45C4-886E-0CC27F8F892A}
 		{D927FB96-2290-4037-8EDB-5D9CA3919169} = {4D21B2A1-8B90-48D3-9D49-7C48B81AE481}
+		{9841C637-B2C4-4927-A6AB-B8CA6869505C} = {4D21B2A1-8B90-48D3-9D49-7C48B81AE481}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7BDAD170-243C-4FCA-95DE-59B8960AD6A8}
