+++ test/PactReferences/CityServiceStates.cs
@@ -1,0 +1,7 @@
+﻿namespace PactReferences;
+
+public static class CityServiceStates
+{
+    public static ProviderStateExtend SomeCitiesExist =>
+        new(Participants.CityService, "SomeCitiesExist", "Some cities exist");
+}
