+++ test/PactReferences/PactReferences.csproj
