+++ test/PactReferences/ProviderStateFunc.cs
@@ -1,3 +1,3 @@
-﻿namespace ProvidersPactStates;
+﻿namespace PactReferences;
 
 public record ProviderStateFunc(string State, Dictionary<string, string> Parameters) { }
