--- test/ProvidersPactStates/CityServiceStates.cs
@@ -1,7 +1,0 @@
-﻿namespace ProvidersPactStates;
-
-public static class CityServiceStates
-{
-    public static ProviderStateExtend SomeCitiesExist =>
-        new(References.CityService, "SomeCitiesExist", "Some cities exist");
-}
